+++ tsconfig.json
@@ -622,10 +622,6 @@
         "ui/styling/style-scope.ts",
         "ui/styling/style.d.ts",
         "ui/styling/style.ts",
-        "ui/styling/stylers-common.ts",
-        "ui/styling/stylers.android.ts",
-        "ui/styling/stylers.d.ts",
-        "ui/styling/stylers.ios.ts",
         "ui/styling/styling.d.ts",
         "ui/styling/styling.ts",
         "ui/styling/visual-state-constants.d.ts",
@@ -639,6 +635,9 @@
         "ui/tab-view/tab-view.android.ts",
         "ui/tab-view/tab-view.d.ts",
         "ui/tab-view/tab-view.ios.ts",
+        "ui/text-base/text-base-styler.android.ts",
+        "ui/text-base/text-base-styler.d.ts",
+        "ui/text-base/text-base-styler.ios.ts",
         "ui/text-base/text-base.d.ts",
         "ui/text-base/text-base.ts",
         "ui/text-field/text-field-common.ts",
