+++ ui/progress/progress.android.ts
@@ -1,6 +1,9 @@
 ﻿import common = require("./progress-common");
 import dependencyObservable = require("ui/core/dependency-observable");
 import proxy = require("ui/core/proxy");
+import styling = require("ui/styling");
+import style = require("ui/styling/style");
+import view = require("ui/core/view");
 
 const R_ATTR_PROGRESS_BAR_STYLE_HORIZONTAL = 0x01010078;
 
@@ -39,3 +42,49 @@ export class Progress extends common.Progress {
         return this._android;
     }
 }
+
+export class ProgressStyler implements style.Styler {
+    private static setColorProperty(view: view.View, newValue: any) {
+        var bar = <android.widget.ProgressBar>view._nativeView;
+        bar.getProgressDrawable().setColorFilter(newValue, android.graphics.PorterDuff.Mode.SRC_IN);
+    }
+
+    private static resetColorProperty(view: view.View, nativeValue: number) {
+        var bar = <android.widget.ProgressBar>view._nativeView;
+        bar.getProgressDrawable().clearColorFilter();
+    }
+
+    private static setBackgroundAndBorderProperty(view: view.View, newValue: any) {
+        var bar = <android.widget.ProgressBar>view._nativeView;
+        var progressDrawable = <android.graphics.drawable.LayerDrawable>bar.getProgressDrawable();
+
+        if (progressDrawable.getNumberOfLayers && progressDrawable.getNumberOfLayers() > 0) {
+            var backgroundDrawable = progressDrawable.getDrawable(0);
+            if (backgroundDrawable) {
+                backgroundDrawable.setColorFilter(newValue, android.graphics.PorterDuff.Mode.SRC_IN);
+            }
+        }
+    }
+
+    private static resetBackgroundAndBorderProperty(view: view.View, nativeValue: number) {
+        var bar = <android.widget.ProgressBar>view._nativeView;
+        // Do nothing.
+    }
+
+    public static registerHandlers() {
+        style.registerHandler(style.colorProperty, new style.StylePropertyChangedHandler(
+            ProgressStyler.setColorProperty,
+            ProgressStyler.resetColorProperty), "Progress");
+
+        style.registerHandler(style.backgroundColorProperty, new style.StylePropertyChangedHandler(
+            ProgressStyler.setBackgroundAndBorderProperty,
+            ProgressStyler.resetBackgroundAndBorderProperty), "Progress");
+
+        style.registerHandler(style.borderWidthProperty, style.ignorePropertyHandler, "Progress");
+        style.registerHandler(style.borderColorProperty, style.ignorePropertyHandler, "Progress");
+        style.registerHandler(style.borderRadiusProperty, style.ignorePropertyHandler, "Progress");
+        style.registerHandler(style.backgroundInternalProperty, style.ignorePropertyHandler, "Progress");
+    }
+}
+
+ProgressStyler.registerHandlers();
