--- ui/styling/stylers.d.ts
@@ -1,3 +1,0 @@
-﻿declare module "ui/styling/stylers" {
-    export function _registerDefaultStylers();
-}
