+++ ui/switch/switch.android.ts
@@ -2,6 +2,9 @@
 import dependencyObservable = require("ui/core/dependency-observable");
 import proxy = require("ui/core/proxy");
 import utils = require("utils/utils")
+import styling = require("ui/styling");
+import style = require("ui/styling/style");
+import view = require("ui/core/view");
 
 function onCheckedPropertyChanged(data: dependencyObservable.PropertyChangeData) {
     var swtch = <Switch>data.object;
@@ -42,3 +45,50 @@ export class Switch extends common.Switch {
         }));
     }
 } 
+
+export class SwitchStyler implements style.Styler {
+    private static setColorProperty(view: view.View, newValue: any) {
+        var sw = <android.widget.Switch>view._nativeView;
+
+        var drawable = <android.graphics.drawable.StateListDrawable>sw.getThumbDrawable();
+        if (drawable) {
+            drawable.setColorFilter(newValue, android.graphics.PorterDuff.Mode.SRC_IN);
+        }
+    }
+
+    private static resetColorProperty(view: view.View, nativeValue: number) {
+        var sw = <android.widget.Switch>view._nativeView;
+        // Do nothing.
+    }
+
+    private static setBackgroundAndBorderProperty(view: view.View, newValue: any) {
+        var sw = <android.widget.Switch>view._nativeView;
+
+        var drawable = <android.graphics.drawable.StateListDrawable>sw.getTrackDrawable();
+        if (drawable) {
+            drawable.setColorFilter(newValue, android.graphics.PorterDuff.Mode.SRC_IN);
+        }
+    }
+
+    private static resetBackgroundAndBorderProperty(view: view.View, nativeValue: number) {
+        var sw = <android.widget.Switch>view._nativeView;
+        // Do nothing.
+    }
+
+    public static registerHandlers() {
+        style.registerHandler(style.colorProperty, new style.StylePropertyChangedHandler(
+            SwitchStyler.setColorProperty,
+            SwitchStyler.resetColorProperty), "Switch");
+
+        style.registerHandler(style.backgroundColorProperty, new style.StylePropertyChangedHandler(
+            SwitchStyler.setBackgroundAndBorderProperty,
+            SwitchStyler.resetBackgroundAndBorderProperty), "Switch");
+
+        style.registerHandler(style.borderWidthProperty, style.ignorePropertyHandler, "Switch");
+        style.registerHandler(style.borderColorProperty, style.ignorePropertyHandler, "Switch");
+        style.registerHandler(style.borderRadiusProperty, style.ignorePropertyHandler, "Switch");
+        style.registerHandler(style.backgroundInternalProperty, style.ignorePropertyHandler, "Switch");
+    }
+}
+
+SwitchStyler.registerHandlers();
