+++ ui/text-base/text-base-styler.d.ts
@@ -1,0 +1,5 @@
+﻿declare module "ui/text-base/text-base-styler" {
+    export class TextBaseStyler {
+        public static registerHandlers();
+    }
+}
