+++ ui/time-picker/time-picker.ios.ts
@@ -1,5 +1,8 @@
 ﻿import common = require("./time-picker-common");
 import dependencyObservable = require("ui/core/dependency-observable");
+import styling = require("ui/styling");
+import style = require("ui/styling/style");
+import view = require("ui/core/view");
 
 function getDate(hour: number, minute: number): NSDate {
     var comps = NSDateComponents.alloc().init();
@@ -92,3 +95,30 @@ class UITimePickerChangeHandlerImpl extends NSObject {
         'valueChanged': { returns: interop.types.void, params: [UIDatePicker] }
     }
 }
+
+export class TimePickerStyler implements style.Styler {
+    // color
+    private static setColorProperty(view: view.View, newValue: any) {
+        var picker = <UIDatePicker>view._nativeView;
+        picker.setValueForKey(newValue, "textColor");
+    }
+
+    private static resetColorProperty(view: view.View, nativeValue: any) {
+        var picker = <UIDatePicker>view._nativeView;
+        picker.setValueForKey(nativeValue, "textColor");
+    }
+
+    private static getColorProperty(view: view.View): any {
+        var picker = <UIDatePicker>view._nativeView;
+        return picker.valueForKey("textColor");
+    }
+
+    public static registerHandlers() {
+        style.registerHandler(style.colorProperty, new style.StylePropertyChangedHandler(
+            TimePickerStyler.setColorProperty,
+            TimePickerStyler.resetColorProperty,
+            TimePickerStyler.getColorProperty), "TimePicker");
+    }
+}
+
+TimePickerStyler.registerHandlers();
